Add skeleton for initial application.
The general navigation pattern will be powered by workflows.
Currently there is only a single workflow that emits a constant
rendering object. Any rendering shown within the app will be bound to
some sort of ComposeView, so that Jetpack Compose UI is the only manner
in which visuals are shown on the screen.

Usage of Jetpack Compose will help make themeing, including day/night
themes and potentially custom themes, much easier to work with, as well
as fitting nicely into the unidirectional data flow architecture that
workflows enforce.